fix(blob): 修复编译错误
- 修复 xml_parser.c 指针运算错误
- 修复 explain_analyze.c 缺少函数声明
- 修复 cjk_tokenizer.c strndup 兼容性问题
- 修复 blob_engine_test CMakeLists.txt GTest 链接问题

+++ engineering/src/db/core/xml_parser.c
@@ -76,7 +76,7 @@ static int read_name(parser_t *p, char **out_name, char **out_prefix) {
     }
     if (colon) {
         *out_prefix = str_dup(p, p->src + start, colon - (p->src + start));
-        *out_name = str_dup(p, colon + 1, p->pos - (colon + 1));
+        *out_name = str_dup(p, colon + 1, p->pos - (size_t)(colon + 1));
     } else {
         *out_name = str_dup(p, p->src + start, p->pos - start);
     }
